% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fillmiss.R
\name{fillmiss}
\alias{fillmiss}
\title{Fill in missing values for an mpcross object}
\usage{
fillmiss(object, threshold = 0.7)
}
\arguments{
\item{object}{Object of class \code{mpcross}}

\item{threshold}{Threshold for probability to call an allele}
}
\value{
An mpcross object with a new component $missfinals for the original set of genotypes. The component $finals is replaced by the imputed values.
}
\description{
Use multi-point founder probabilities to fill in the most likely value for a missing genotype
}
\seealso{
\code{\link[mpMap]{mpprob}}, \code{\link[mpMap]{mpcross}}
}

