% Generated by roxygen2 (4.0.2): do not edit by hand
\name{findqtl}
\alias{findqtl}
\title{Detect a QTL peaks in a QTL profile from (composite) interval mapping}
\usage{
findqtl(mpqtl, dwindow = 5, threshold)
}
\arguments{
\item{mpqtl}{Object of class \code{mpqtl}}

\item{dwindow}{Window over which to smooth p-values - default is five markers}

\item{threshold}{Threshold peaks must exceed to be detected (-log10(p))}
}
\value{
The original input object with additional entries for newly detected QTL.
}
\description{
Given the output from a scan of a chromosome, locates local maxima exceeding a significance threshold in a QTL profile.
}
\examples{
sim.map <- sim.map(len=rep(100, 2), n.mar=11, include.x=FALSE, eq.spacing=TRUE)
sim.ped <- sim.mpped(4, 1, 500, 6, 1)
sim.dat <- sim.mpcross(map=sim.map, pedigree=sim.ped, qtl=matrix(data=c(1, 10, .4, 0, 0, 0, 1, 70, 0, .35, 0, 0), nrow=2, ncol=6, byrow=TRUE), seed=1)
mpp.dat <- mpprob(sim.dat, program="qtl", step=2)
mpq.dat <- mpIM(object=mpp.dat, ncov=0, responsename="pheno")
mpq2 <- findqtl(mpq.dat, dwindow=5, threshold=3)
plot(mpq2)
summary(mpq2)
}
\seealso{
\code{\link[mpMap]{mpIM}}, \code{\link[mpMap]{plot.mpqtl}}, \code{\link[mpMap]{summary.mpqtl}}
}

