\name{get.binmarkers}
\alias{get.binmarkers}
\title{Get Bin Indexes}
\usage{
  get.binmarkers(object, chr, bin)
}
\arguments{
  \item{object}{Object of class \code{mpcross}}

  \item{chr}{The chromosome where the bin is located}

  \item{bin}{The bin to get marker indexes for}
}
\value{
  A vector giving the indexes of all markers in the
  specified location
}
\description{
  Returns the indexes of of markers in a particular
  chromosome and bin
}

