% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/identifyDesign.R
\name{identifyDesign}
\alias{identifyDesign}
\title{Identify subtype of design in pedigree}
\usage{
identifyDesign(pedigree)
}
\arguments{
\item{pedigree}{Pedigree to categorize}
}
\value{
An additional vector identifying designs of individuals within the pedigree
}
\description{
For a pedigree containing multiple variations on MAGIC designs, identify
the subtype of design, e.g., MAGIC+
}

