% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maporder.R
\name{maporder}
\alias{maporder}
\title{Helper function to ensure that mpcross objects are in map order}
\usage{
maporder(object)
}
\arguments{
\item{object}{Object of class \code{mpcross}}
}
\value{
The original object, with all components reordered to match the map. Additional unmapped markers will appear after all mapped markers.
}
\description{
Orders markers within all components of mpcross object (founders, finals, etc.) to match the map order
}

