% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp-mapdist.R
\name{haldaneR2X}
\alias{haldaneR2X}
\alias{haldaneX2R}
\alias{kosambiR2X}
\alias{kosambiX2R}
\alias{mp.mapdist}
\title{Conversion between recombination fractions (R) and map distance (X)}
\usage{
haldaneR2X(r)

haldaneX2R(x)

kosambiR2X(r)

kosambiX2R(x)
}
\arguments{
\item{r}{Recombination fraction}

\item{x}{Map distance (measured in centiMorgans)}
}
\description{
Haldane and Kosambi map functions and inverses for converting between 
recombination fractions and map distance (in cM)
}
\references{
Jurg Ott, Analysis of Human Genetic Linkage, Johns Hopkins University Press, Baltimore 1999
}

