% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mpclean}
\alias{mpclean}
\title{Find Largest Complete mpcross object}
\usage{
mpclean(object, output.type = c("summary", "object"))
}
\arguments{
\item{object}{Object of class \code{mpcross}}

\item{output.type}{Type of output}
}
\value{
If the output type is "summary",  then a 2 x n.chr matrix is returned. The first row gives the number of markers present on that linkage group. The second row gives the number of markers in the maximal subset. If the output type is "object" then an \code{mpcross} object is returned. The object is a subset of the orginial object, with the largest set of markers where all pairs of recombination fractions estimates are availbe in each linkage group.
}
\description{
Finds the largest subset of markers for which all recombination fraction estimates are available for each linkage group. The subsetted object can be returned, or a summary of marker reduction can be given.
}
\examples{
map <- sim.map(len=1, n.mar=30, eq.spacing=FALSE, include.x=FALSE)
sim.ped <- sim.mpped(4, 1, 500, 6, 1)
sim.dat <- sim.mpcross(map=map, pedigree=sim.ped, seed=1)
dat.rf <- mpestrf(sim.dat)
dat.lg<-mpgroup(dat.rf, groups=1)
clean.summary <- mpclean(dat.lg)
}

