% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mpIMlm.R
\name{mpIMlm}
\alias{mpIMlm}
\title{(Composite) Interval Mapping with linear models for QTL detection in multi-parent crosses}
\usage{
mpIMlm(object, threshold = 0.001, chr, step = 0, responsename = "predmn",
  ncov = 1000, window = 10, dwindow = 5, mrkpos = TRUE, fixed,
  foundergroups, ...)
}
\arguments{
\item{object}{Object of class \code{mpcross}}

\item{threshold}{Significance threshold for QTL p-values}

\item{chr}{Subset of chromosomes for which to compute QTL profile}

\item{step}{Step size at which to compute the QTL profile. See \code{\link[mpMap]{mpprob}} for further description of default values}

\item{responsename}{Response name for testing}

\item{ncov}{Number of marker covariates to search for - default is to search for as many as possible using stepAIC (forward/backward selection)}

\item{window}{Window of cM on each side of markers where we exclude covariates in CIM}

\item{dwindow}{Window of markers to use for smoothing in QTL detection}

\item{mrkpos}{Flag for whether to consider both marker positions and step positions or just steps. Is overridden if step=0}

\item{fixed}{If input, vector of fixed effects for each individual to be included in model with main effect and interaction with founder probability}

\item{foundergroups}{If input, groups of founders for which to cluster parental alleles together at every marker. Currently overrides mrkpos and step arguments. Note that this is not currently working with ncov>0}

\item{...}{Additional arguments}
}
\value{
The original input object with additional component QTLresults containing the following elements:
\item{pheno}{Input phenotype data}
\item{pvalue}{Each component contains estimated p-values at each position on a given chromosome}
\item{wald}{Each component contains Wald statistics at each position on a given chromosome}
\item{fndrfx}{Each component contains founder effects estimated at each position on a given chromosome}
\item{qtl}{Each component contains the position and effects of a detected QTL}
\item{fixedmain}{Each component contains wald statistics for main effect of fixed variable (if input) at each position on a given chromosome}
\item{fixedintx}{Each component contains wald statistics at each position on a given chromosome for gene x fixed interaction (if input)}
\item{fixedintdf}{Each component contains the df for the gene x fixed interaction at each position on a given chromosome}
\item{call}{Input arguments to function}
and with attributes describing the number of QTL detected, and the threshold used for detection. Note: Now uses the function findqtl to find all QTL peaks, see \code{\link[mpMap]{findqtl}} for more information.
}
\description{
Interval mapping in multi-parent crosses using multi-stage linear model
approach including ability to use cofactors (CIM)
}
\details{
Modified version of mpIM to just fit linear models for QTL mapping.
This has been separated off from the main function as some functionality
has been recently implemented just for the linear model fitting - this
allows for rapid development of the linear model functionality while
maintaining the mixed model functionality in a stable state more easily.
No argument names have been changed, hence mpIMlm will be called in the
same way as the previous version of mpIM. The only requirement is that
phenotypes are included in the mpcross object rather than allowing
for separate input of a phenotype matrix.
Note that no weights are used in this analysis, which may result in a loss of efficiency compared to a single-stage approach.

If fixed is input will add terms to the model to test for a fixed effect of
the input vector (so make sure the class is correct) and for an interaction
between the input vector and the founder haplotypes. Note that only a single
fixed covariate can currently be included to avoid overparametrization.

If foundergroups is input, then probabilities at each location will be collapsed within the groups of founders
in fitting the model.
}
\examples{
sim.map <- sim.map(len=rep(100, 2), n.mar=11, include.x=FALSE, eq.spacing=TRUE)
sim.ped <- sim.mpped(4, 1, 500, 6, 1)
sim.dat <- sim.mpcross(map=sim.map, pedigree=sim.ped, qtl=matrix(data=c(1, 10, .4, 0, 0, 0, 1, 70, 0, .35, 0, 0), nrow=2, ncol=6, byrow=TRUE), seed=1)
mpp.dat <- mpprob(sim.dat, program="qtl")
## Two-stage simple interval mapping
mpq.dat <- mpIM(object=mpp.dat, ncov=0, responsename="pheno")
}
\seealso{
\code{\link[mpMap]{plot.mpqtl}}, \code{\link[mpMap]{summary.mpqtl}}, \code{link[mpMap]{fit.mpqtl}}
}

