\name{mpMap-package}
\Rdversion{1.1}
\alias{mpMap-package}
\alias{mpMap}
\docType{package}
\title{
Genetic analysis in multi-parent crosses
}
\description{
Simulates data from MAGIC designs (4/8 parents), constructs linkage maps, 
reconstructs haplotypes, and performs QTL mapping
}
\details{
\tabular{ll}{
Package: \tab mpMap\cr
Type: \tab Package\cr
Version: \tab 2.0.2\cr
Date: \tab 2015-07-29\cr
License: \tab GPL-2 \cr
LazyLoad: \tab yes\cr
}
Functions for data manipulation: \code{\link[mpMap]{mpcross}}, \code{\link[mpMap]{read.mpcross}}, \code{\link[mpMap]{write.mpcross}}, \code{\link[mpMap]{cleanmap}}, \code{\link[mpMap]{clean.mpcross}}, \code{\link[mpMap]{mpadd}}, \code{\link[mpMap]{subset.mpcross}} \cr
Functions for simulation: \code{\link[mpMap]{sim.mpcross}}, \code{\link[mpMap]{sim.mpped}} \cr
Functions for map construction: \code{\link[mpMap]{mpestrf}}, \code{\link[mpMap]{mpgroup}}, \code{\link[mpMap]{mporder}}, \code{\link[mpMap]{add3pt}}, \code{\link[mpMap]{compare_orders}} \cr
Functions for haplotype reconstruction: \code{\link[mpMap]{mpprob}} \cr
Functions for QTL mapping: \code{\link[mpMap]{mpIM}}, \code{\link[mpMap]{sim.sigthr}}, \code{\link[mpMap]{findqtl}}
}
\author{
Emma Huang and Rohan Shah

Maintainer: Emma Huang <Emma.Huang@csiro.au>
}
\keyword{ package }
