% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phys2gen.R
\name{phys2gen}
\alias{phys2gen}
\title{Estimate a genetic map based on physical map ordering in multi-parent cross}
\usage{
phys2gen(physmap, mpcross, mapfx, maxrf = 0.25)
}
\arguments{
\item{physmap}{Physical map to be used for groups and orders of markers}

\item{mpcross}{Mpcross objects from which to estimate recombination fractions}

\item{mapfx}{Map function to convert recombination fractions to cM}

\item{maxrf}{Maximum rf to allow between markers before removing from map}
}
\value{
A pgmap object is returned containing both genetic and physical map positions.
}
\description{
Use an input physical map to group and order markers; estimate positions
based on recombination fractions between adjacent markers
}
\details{
For each pair of adjacent markers recombination fractions are estimated,
and the map positions computed as the sum of adjacent cM values. The
returned object is of class "pgmap", which contains both physical and
genetic map information for ease of later comparison.
}

