% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlink.map.R
\name{plotlink.map}
\alias{plotlink.map}
\title{Plots linkage maps}
\usage{
plotlink.map(object, chr, marker.names = TRUE, colqtl = "red", ...)
}
\arguments{
\item{object}{Either \code{mpcross} or \code{map} object}

\item{chr}{Chromosomes to plot}

\item{marker.names}{Whether to plot marker names}

\item{colqtl}{Color to plot QTL regions.}

\item{...}{Additional arguments}
}
\value{
Modification of link.map.cross function from wgaim to allow more
general input objects and to highlight regions around QTL. If any markers
are labelled "QTLx" then they will be plotted in a different color.
}
\description{
Plot linkage map (either as input object or as stored in mpcross object). 
Can also highlight QTL regions when used with qtlmap function. Modification of
code from package wgaim.
}
\author{
Emma Huang, Julian Taylor
}

