% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compare-orders.R
\name{compare_orders}
\alias{compare_orders}
\title{Compare potential orders for linkage groups}
\usage{
compare_orders(cross, chr, orders, method = c("countXO", "likelihood"))
}
\arguments{
\item{cross}{Object of class \code{mpcross} or \code{cross}}

\item{chr}{Selected chromosomes}

\item{orders}{Orders to be compared}

\item{method}{Method for comparison. Note that "likelihood" is much more time-consuming than "countXO"}
}
\value{
The matrix of orders with crossover counts appended.
}
\description{
Compares potential orderings on the number of expected crossovers or
likelihood values. Rewrite of older version.
}
\details{
Uses functions from R/qtl in order to compare possible orderings on the basis of the number of obligate crossovers or log-likelihood values.
}
\examples{
map <- qtl::sim.map(len=100, n.mar=11, eq.spacing=TRUE, include.x=FALSE)
sim.ped <- sim.mpped(4, 1, 500, 6, 1)
sim.dat <- sim.mpcross(map=map, pedigree=sim.ped, seed=1)
compare_orders(sim.dat, chr=1, orders=rbind(1:11, c(1:3, 6:4, 7:11)))
}
\references{
R/qtl
}
\seealso{
\code{\link[mpMap]{mporder}}, \code{\link[qtl]{countXO}}
}

