% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mpcollapse.R
\name{mpcollapse}
\alias{mpcollapse}
\title{Collapse an mpcross object into a binned object}
\usage{
mpcollapse(object, method = c("rf", "cor"), adj = FALSE, cutoff = 0,
  consensusProbability = 0.75, missingCutoff = 0.25)
}
\arguments{
\item{object}{Object of class \code{mpcross}}

\item{method}{Choice of whether to bin based on recombination fraction or correlation}

\item{adj}{Flag for whether require bins to be based on adjacent markers only}

\item{cutoff}{Max RF allowed before starting a new bin}

\item{consensusProbability}{Proportion of genotypes within FDP groups which must match for imputation to occur.}

\item{missingCutoff}{Proportion of missing data allowed within markers before not using to create haplotypes}
}
\value{
A binned object of class \code{binmpcross}
}
\description{
Collapses an mpcross object into bins. Within bins, markers are grouped by founder distribution pattern and missing genotypes imputed. Haplotypes are formed for markers with unique FDPs and determine the bin alleles. Information on which markers belong to each bin is retained within the output object.
}
\note{
Markers are assigned to a bin if there is zero recombination between any of the markers within the bin (or correlation of 1). Next, a three-step process is performed to create the new set of bin markers. First, markers within a bin are grouped based on founder distribution pattern (FDP). Second, for each FDP, a single representative marker is chosen, with missing values imputed from all markers sharing that FDP. Third, haplotypes formed from these representative markers are used to replace genotypes for the bin marker.
}
\seealso{
\code{link{mpexpand}}
}

